feat: extract ProjectDemandChart data processing to reusable hooks (closes #63)
- Create useProjectDemandData hook for demand/capacity/gaps data processing
- Create useChartGranularity hook for granularity calculation and aggregation
- Create useChartDisplayData hook for brush selection filtering
- Reduce ProjectDemandChart.tsx from 1,654 to 892 lines (~46% reduction)
- Move 649 lines of data processing logic to dedicated hooks
- Hooks can be reused by other chart components
